ERTool libraries and formula update
ERTool Updates:
- Fix #14, log4j update to 2.19
- SWT library was updated in conjunction to support the new log4j
- CharWidth to manipulate pixel size in erGUI due to SWT update

Analytics.xml updates
- End of service updates reflected (v2018, v10.0.1)
- End of hardware updates reflected (8436)
- 10.0.2-10.0.4 now show version "orange" in top title as warning since 10.0.2/10.0.3 are not supported, 10.0.4.0-sr3 while supported, does not have latest fixes (sr2/sr1 out of date)
- 10.0.1 now shows version as orange in top title since out of service April 2024
- Default log, before reload and queue log are listed as a file section
- CLILog formula now provides 'beforereload' files as well, no longer tries to decode cli logs (parameter was accidentally set)
- DPOSDiagnostics is now setup to auto decode (alternate SysRecord)
- Restricted the interfaces found hardware error to only display if we have LESS interfaces, eg 16/18.  If you have 19/18 it will ignore.
- Table based display on the GatewayPeeringStatus added in 10.5
- Some old formulas deprecated (they will only run against v2018 since there are new status providers in v10 to replace them)

+++ framework/src/com/ibm/datapower/er/Analytics/AnalyticsProcessor.java
@@ -1573,7 +1573,13 @@ public class AnalyticsProcessor {
 						File endFile = new File(dir + endFileName);
 						boolean exists = endFile.exists();
 						if(decode && exists) {
-							endFile.delete();
+							File undecodedFileName = new File(dir + endFileName + ".undecoded");
+							if(!undecodedFileName.exists()) {
+								endFile.renameTo(undecodedFileName);
+							}
+							else {
+								endFile.delete();
+							}
 							exists = false;
 						}
 						if (!exists) {
